backend/requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
motor==3.7.1
pymongo==4.16.0
pydantic==2.12.5
pydantic-settings==2.7.0
email-validator==2.2.0
python-jose==3.5.0
passlib==1.7.4
bcrypt==4.2.1
python-multipart==0.0.22
python-dotenv==1.2.2
openai==1.58.1
httpx==0.28.1
anyio==4.12.1
starlette==0.52.1
certifi==2026.2.25
dnspython==2.8.0
idna==3.11
h11==0.16.0
annotated-types==0.7.0
typing-extensions==4.15.0
click==8.3.1
six==1.17.0
rsa==4.9.1
pyasn1==0.6.2
ecdsa==0.19.1
